+++ src/routes/ConsortiumManager/settings/authorization-policies/constants.js
@@ -1,9 +1,6 @@
-import { FormattedDate, FormattedMessage } from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 
-import { NoValue, TextLink } from '@folio/stripes/components';
-import { getFullName } from '@folio/stripes/util';
-
-const COLUMN_NAMES = {
+export const COLUMN_NAMES = {
   name: 'name',
   description: 'description',
   updated: 'updated',
@@ -23,15 +20,3 @@ export const VISIBLE_COLUMNS = [
   COLUMN_NAMES.updated,
   COLUMN_NAMES.updatedBy,
 ];
-
-export const getResultsFormatter = ({ users }) => ({
-  [COLUMN_NAMES.name]: (item) => <TextLink>{item.name}</TextLink>,
-  [COLUMN_NAMES.updatedBy]: (item) => (item.metadata.updatedByUserId
-    ? getFullName(users[item.metadata.updatedByUserId])
-    : <NoValue />
-  ),
-  [COLUMN_NAMES.updated]: (item) => (item.metadata.updatedDate
-    ? <FormattedDate value={item.metadata.updatedDate} />
-    : <NoValue />
-  ),
-});
